# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CopyPasteHouse API",
        "description": "API documentation for CopyPasteHouse - A paste sharing application",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/syntax-highlights": {
            "get": {
                "tags": [
                    "Syntax Highlights"
                ],
                "summary": "List all syntax highlights",
                "description": "Retrieve a list of available syntax highlighting options",
                "operationId": "6434d8c5db69cb7c1f4fa611b05713ce",
                "responses": {
                    "200": {
                        "description": "List of syntax highlights",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "code": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Syntax Highlights"
                ],
                "summary": "Create a syntax highlight",
                "description": "Create a new syntax highlighting option",
                "operationId": "4fbb0aa8763fd186ead1bc15e5950a19",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "code"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "JavaScript"
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "javascript"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Syntax highlight created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "syntax_highlight": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Syntax highlight created successfully!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/syntax-highlights/{syntaxHighlight}": {
            "put": {
                "tags": [
                    "Syntax Highlights"
                ],
                "summary": "Update a syntax highlight",
                "description": "Update an existing syntax highlighting option",
                "operationId": "c300473ebe0d20c07c0f72c607f7d129",
                "parameters": [
                    {
                        "name": "syntaxHighlight",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Python"
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "python"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Syntax highlight updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "syntax_highlight": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Syntax highlight updated successfully!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Syntax highlight not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Syntax Highlights"
                ],
                "summary": "Delete a syntax highlight",
                "description": "Delete an existing syntax highlighting option",
                "operationId": "1ad3884e5c902e4f9154e156ad2e40da",
                "parameters": [
                    {
                        "name": "syntaxHighlight",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Syntax highlight deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Syntax highlight deleted successfully!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Syntax highlight not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your bearer token in the format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}